+++ Tortuga.Chain/Tortuga.Chain.SqlServer.MDS/SqlServer/CommandBuilders/ISupportsTableHints.cs
@@ -1,0 +1,11 @@
+namespace Tortuga.Chain.SqlServer.CommandBuilders;
+
+interface ISupportsTableHints
+{
+	void AddTableHint(string hint);
+}
+
+interface ISupportsQueryHints
+{
+	void AddQueryHint(string hint);
+}